+++ src/CustomBinaryOutputPort.cpp
@@ -74,7 +74,7 @@ int CustomBinaryOutputPort::putU8(uint8_t v)
     const Object count = Object::makeFixnum(1);
     const Object result = theVM_->callClosure3(writeDProc_, bv, start, count);
     MOSH_ASSERT(result.isFixnum());
-    return result.toFixnum();
+    return static_cast<int>(result.toFixnum());
 }
 
 int64_t CustomBinaryOutputPort::putU8(uint8_t* v, int64_t _size)
